# refine build strategy for the model optimize test=develop