using PR https://github.com/PaddlePaddle/Paddle/pull/19738 whl package
average latency of 151 steps, skip 0 step:
	Avg: 0.245 s/step
	FPS: 8.159 samples/s
average latency of 151 steps, skip 2 steps:
	Avg: 0.245 s/step
	Min: 0.222 s/step
	Max: 0.273 s/step
	FPS: 8.162 samples/s